Refactor | navBar 리팩토링에 의한 navbarProps 수정
